refactor(ios): extract schema statics to DatabaseManager+Schema; fix file_length
The pre-migration backup hook pushed DatabaseManager.swift past the 1000-line
file_length limit (error severity, fails CI). Move the pre-migration backup
orchestration into BackupService (backUpForPendingMigrations), and split the
v25 schema/index/sync-table creation statics out into DatabaseManager+Schema.swift.
DatabaseManager.swift drops to ~540 lines; no behavior change.
